# Improved loading performance when using cached address spaces