Hey,

I've tried to run the python-opcua server on a Raspberry-Pi.
Unfortunately starting the server and especially the construction of the address spaces takes quite a while (~57secs), so I started to play with the already built in caching of the address spaces.

To my surprise the loading performance was much worse when loading the the address space from a pickle file (~157secs).

However, simply working with the C-Implementation of the pickle mechanism and using the latest protocol version is delivering better results (cPython: ~33 secs / cPython + HIGHEST_PROTOCOL: ~22 secs).

I didn't do any exact performance tests so you shouldn't take those values too seriously :-)
However, since other persons report similar benchmark results (e.g. http://www.benfrederickson.com/dont-pickle-your-data/), I guess it might also be a performance improvement when running on other platforms...