@oroulet 2016-01-08
Since you are looking at performance, how does pypy works on rasperry? And how fast is it to load address space from xml instead of code or pickel?