@stanti 2016-01-08
The results above were created with python 2.7.
Python3 delivered the fasted import (13secs from pickle / 70 secs from generated code).
Pypy (2.7 compatible) was slower than cpython (93 secs from pickle / 80 secs from generated code).

I never checked the xml import since it didn't finish without errors and I didn't expect it to be faster than pickle/generated code.